requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
crispy-bootstrap4==2024.10
cssselect2==0.7.0
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.2
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==24.3
django-mathfilters==1.0.0
docopt==0.6.2
fonttools==4.54.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Markdown==3.8.2
markdown-it-py==3.0.0
-e git+https://github.com/modelcontextprotocol/python-sdk.git@6f43d1f17022d9d45b51955304e39ad43ca48014#egg=mcp
mdurl==0.1.2
num2words==0.5.13
openai==1.93.0
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
pyphen==0.16.0
python-dotenv==1.0.1
python-multipart==0.0.20
referencing==0.36.2
rich==14.0.0
rpds-py==0.25.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
sse-starlette==2.3.6
starlette==0.47.1
tinycss2==1.3.0
tinyhtml5==2.0.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
uvicorn==0.35.0
weasyprint==62.3
webencodings==0.5.1
whitenoise==6.8.2
zopfli==0.2.3
